requirements.txt
GitPython==3.1.30
PyDriller==2.5
networkx==3.2.1
python-louvain==0.16
numpy==1.26.2
pandas==2.1.3

# FastAPI stack (per Week 1)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
neo4j==5.14.1
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.6

# ORM & tooling
SQLAlchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-dotenv==1.0.0
